+++ src/panels/importSourcesOnlyResultPanel.ts
@@ -87,7 +87,7 @@ export class ImportSourcesOnlyResultPanel {
    * are created and inserted into the webview HTML.
    *
    * @param extensionUri The URI of the directory containing the extension
-   * @returns A template string literal containing the HTML that should be
+   * @returns A _template_ string literal containing the HTML that should be
    * rendered within the webview panel
    */
   private _getWebviewContent(extensionUri: vscode.Uri) {