+++ src/panels/inspectObjectPanel.ts
@@ -116,7 +116,7 @@ export class InspectorObjectPanel extends TdsPanel<TInspectorObjectModel, IInspe
    * are created and inserted into the webview HTML.
    *
    * @param extensionUri The URI of the directory containing the extension
-   * @returns A template string literal containing the HTML that should be
+   * @returns A _template_ string literal containing the HTML that should be
    * rendered within the webview panel
    */
   protected getWebviewContent(extensionUri: vscode.Uri) {