+++ src/panels/panel.ts
@@ -68,7 +68,7 @@ export abstract class TdsPanel<M extends TAbstractModelPanel, O extends any = {}
 	 * are created and inserted into the webview HTML.
 	 *
 	 * @param extensionUri The URI of the directory containing the extension
-	 * @returns A template string literal containing the HTML that should be
+	 * @returns A _template_ string literal containing the HTML that should be
 	 * rendered within the webview panel
 	 */
 	protected abstract getWebviewContent(extensionUri: vscode.Uri): string;