+++ src/protocolMessages.ts
@@ -594,46 +594,6 @@ export function sendPatchInfo(
     );
 }
 
-export interface IApplyTemplateResult {
-  error: boolean;
-  message: string;
-  errorCode: number;
-}
-
-export function sendApplyTemplateRequest(
-  server: ServerItem,
-  includesUris: Array<string>,
-  templateUri: vscode.Uri
-): Thenable<IApplyTemplateResult> {
-  return languageClient
-    .sendRequest("$totvsserver/templateApply", {
-      templateApplyInfo: {
-        connectionToken: server.token,
-        authorizationToken: ServersConfig.getAuthorizationToken(server),
-        environment: server.environment,
-        includeUris: includesUris,
-        templateUri: templateUri.toString(),
-        isLocal: true,
-      },
-    })
-    .then(
-      (response: IApplyTemplateResult) => {
-        if (response.error) {
-          return Promise.reject(response);
-        }
-        return Promise.resolve(response);
-      },
-      (err: ResponseError<object>) => {
-        const error: IApplyTemplateResult = {
-          error: true,
-          message: err.message,
-          errorCode: err.code,
-        };
-        return Promise.reject(error);
-      }
-    );
-}
-
 interface IRpoTokenResult {
   sucess: boolean;
   message: string;